+++ src/SauceTool.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
-#include "Sauce.h" 
+#include "SauceTool.h" 
 
 // Compiler and OS defines
