+++ test/src/CheckTest.c
@@ -1,5 +1,5 @@
 #include "unity.h"
-#include "Sauce.h"
+#include "SauceTool.h"
 #include "TestRes.h"
 #include <string.h>
 #include <stdio.h>
